+++ plugin-maven/CHANGELOG.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- `equo-ide:launch -Dclean=true` now launches a totally fresh IDE. ([#50](https://github.com/equodev/equo-ide/pull/50))
 
 ## [0.9.0] - 2023-01-09
 ### Added